+++ src/main/java/com/penguineering/mnrmapi/index/IndexAccess.java
@@ -1,6 +1,6 @@
 package com.penguineering.mnrmapi.index;
 
-import com.penguineering.mnrmapi.gcs.GcsAccess;
+import com.penguineering.mnrmapi.GCS;
 import io.micronaut.context.annotation.Bean;
 import jakarta.inject.Inject;
 import reactor.core.publisher.Flux;
@@ -9,11 +9,11 @@ import reactor.core.publisher.Mono;
 @Bean
 public class IndexAccess {
     @Inject
-    protected GcsAccess gcsAccess;
+    protected GCS gcs;
 
     public Flux<IndexEntry> retrieveIndex(Flux<String> gcsPath) {
         return gcsPath
-                .transform(gcsAccess::retrieve)
+                .transform(gcs::retrieve)
                 .map(IndexEntry::fromData)
                 .flatMap(Flux::fromIterable);
     }
@@ -21,7 +21,7 @@ public class IndexAccess {
     public Flux<String> retrieveRootGcs() {
         return Flux
                 .just("root")
-                .transform(gcsAccess::retrieve);
+                .transform(gcs::retrieve);
     }
 
     public Flux<IndexEntry> retrieveRootIndex() {
